creates BodyareaLifts controller and moves getLiftsByBodyarea logic out of the server and into the controller
